Moved servers into their own package
