+++ node.py
@@ -1,82 +1,8 @@
 from miner import Miner
-from protos import request_pb2
-from google.protobuf import message
-import time
-import server
+from servers import server
+from servers.data_server import DataServer
+from servers.request_server import RequestServer
 import logging
-import peer_to_peer_discovery as p2p
-from requests import RequestParser
-
-
-class DataServer(server.TCPRequestHandler):
-    """
-    The data server for receiving incoming TCP binary data from outside the peer to peer network.
-    """
-
-    def receive(self, data):
-        """
-        Receive binary data from an incoming TCP request that should be added to the block chain.
-        :param data: The binary data to be added to the block chain.
-        """
-        request = request_pb2.BlobMessage()
-        request.timestamp = time.time()
-        request.blob = data
-
-        msg = request.SerializeToString()
-        logging.debug("Received data: (%f, %s) = %s", request.timestamp, request.blob, msg)
-        self.server.miner.add(msg)
-
-        # TODO forward msg to peers
-
-
-class RequestServer(server.TCPRequestHandler):
-    """
-    The request server for handling requests from nodes in the peer to peer network.
-    """
-
-    def __init__(self, request, client_address, serv):
-        self.parser = RequestParser(self)
-        server.TCPRequestHandler.__init__(self, request, client_address, serv)
-
-    def receive(self, data):
-        """
-        Received binary messages from other nodes in the peer to peer network.
-        :param data: The binary data which should be decodable using the Request protocol buffer
-        """
-        self.parser.parse(data)
-
-    def handle_blob(self, data):
-        pass
-
-    def handle_alive(self, data):
-        pass
-
-    def handle_mined_block(self, data):
-
-        msg = request_pb2.MinedBlockMessage()
-        try:
-            msg.ParseFromString(data)
-        except message.DecodeError:
-            logging.error("Error decoding message: %s", data)
-            return
-
-        self.server.miner.receive_block(msg.block, msg.chain_cost)
-
-
-class DiscoveryServer(server.UDPRequestHandler):
-
-    def __init__(self, request, client_address, serv):
-        self.nodes = []
-
-        server.UDPRequestHandler.__init__(self, request, client_address, serv)
-
-    def receive(self, data):
-
-        logging.debug("Got broadcast message")
-
-        if self.client_address[0] not in self.nodes:
-            self.nodes.append(self.client_address[0])
-            logging.debug("Received new ip %s", str(self.client_address[0]))
 
 
 class Node:
@@ -94,8 +20,8 @@ class Node:
         self.input_server = server.TCPServer(9999, DataServer)
         self.input_server.miner = self.miner
 
-        self.udp_server = server.UDPServer(10005, DiscoveryServer)
-        self.udp_broadcaster = p2p.UDPBroadcaster(10005, 5)
+        # self.udp_server = server.UDPServer(10028, DiscoveryServer)
+        # self.udp_broadcaster = p2p.UDPBroadcaster(10028, 5)
 
     def block_mined(self, block, chain_cost):
         pass
@@ -105,9 +31,11 @@ class Node:
         Run the servers for receiving incoming requests and start mining.
         This method never returns.
         """
+        logging.debug("enable")
+
         server.start_server(self.request_server)
         server.start_server(self.input_server)
-        server.start_server(self.udp_server)
-        p2p.start_discovery(self.udp_broadcaster)
+        # server.start_server(self.udp_server)
+        # p2p.start_discovery(self.udp_broadcaster)
 
         self.miner.mine()
