+++ peer_to_peer_discovery.py
@@ -27,6 +27,7 @@ class UDPBroadcaster:
         """
         sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
         sock.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, 1)
+        sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
 
         # Send a message
         message = disc_msg.DiscoveryMessage()
